integering evaluation and intiialization time for RTDP

+++ solvers/rtdp/rtdp.cpp
@@ -106,7 +106,7 @@ void RtdpPolicy::train() {
     auto elapsed_time_milliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(
             init_end - init_begin).count();
     float elapsed_time_seconds = float(elapsed_time_milliseconds) / 1000;
-    (*(this->train_info->additional_data))["init_time"] = std::to_string(round(elapsed_time_seconds * 100) / 100);
+    (*(this->train_info->additional_data))["init_time"] = std::to_string((int)round(elapsed_time_seconds * 100) / 100);
     bool converged = false;
     size_t iters_count = 0;
     EvaluationInfo *prev_eval_info = NULL;
@@ -145,7 +145,7 @@ void RtdpPolicy::train() {
     elapsed_time_seconds = float(elapsed_time_milliseconds) / 1000;
     total_eval_time = float(total_eval_time) / 1000;
     this->train_info->time = round(elapsed_time_seconds * 100) / 100;
-    (*(this->train_info->additional_data))["eval_time"] = std::to_string(round(total_eval_time * 100) / 100);
+    (*(this->train_info->additional_data))["eval_time"] = std::to_string((int)round(total_eval_time * 100) / 100);
     (*(this->train_info->additional_data))["n_iterations"] = std::to_string(iters_count + 1);
 
     /* Finish training */
